refactor(mock): mock data 위치 및 폴더구조 변경
