+++ src/mocks/browser.ts
@@ -1,11 +1,11 @@
 import { setupWorker } from 'msw';
 
-import { categoriesMockHandler } from '@/api/hooks/categories.mock';
-import { loginMockHandler } from '@/api/hooks/login.mock';
-import { productOptionsMockHandler } from '@/api/hooks/product-options.mock';
-import { productsMockHandler } from '@/api/hooks/products.mock';
-import { registerMockHandler } from '@/api/hooks/register.mock';
-import { getWishListMockHandler } from '@/api/hooks/wishList.mock';
+import { loginMockHandler } from './auth/login.mock';
+import { registerMockHandler } from './auth/register.mock';
+import { categoriesMockHandler } from './categories/categories.mock';
+import { productOptionsMockHandler } from './product/product-options.mock';
+import { productsMockHandler } from './product/products.mock';
+import { getWishListMockHandler } from './wish-list/wishList.mock';
 
 export const worker = setupWorker(
   ...categoriesMockHandler,
